Add mechanism for generating a header with no output
For callers interested only in the chunk/hash map, allow
them to pass an invalid fd for output. When that occurs,
do not create a temporary file, and write no output at all.

This also reworks the temp_fd checks slightly to allow for
the possibility that mkstemp() returns an fd of 0, which is
valid (though unlikely) on UNIX systems.

Signed-off-by: Matt Madison <matt@madison.systems>